Math.toIntExact is unnecessary because value is guaranteed to be between 0 and 15
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1915959 13f79535-47bb-0310-9956-ffa450edef68

+++ poi/src/main/java/org/apache/poi/util/HexDump.java
@@ -294,7 +294,7 @@ public final class HexDump {
         char[] buf = new char[nDigits];
         long acc = value;
         for(int i=nDigits-1; i>=0; i--) {
-            int digit = Math.toIntExact(acc & 0x0F);
+            int digit = (int) (acc & 0x0F);
             buf[i] = (char) (digit < 10 ? ('0' + digit) : ('A' + digit - 10));
             acc >>>= 4;
         }
